gvcp: use guint16 for packet_count type.

+++ src/arvgvcp.c
@@ -48,7 +48,7 @@ arv_gvcp_packet_free (ArvGvcpPacket *packet)
  */
 
 ArvGvcpPacket *
-arv_gvcp_packet_new_read_memory_cmd (guint32 address, guint32 size, guint32 packet_count, size_t *packet_size)
+arv_gvcp_packet_new_read_memory_cmd (guint32 address, guint32 size, guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -83,7 +83,7 @@ arv_gvcp_packet_new_read_memory_cmd (guint32 address, guint32 size, guint32 pack
  */
 
 ArvGvcpPacket *
-arv_gvcp_packet_new_read_memory_ack (guint32 address, guint32 size, guint32 packet_count, size_t *packet_size)
+arv_gvcp_packet_new_read_memory_ack (guint32 address, guint32 size, guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -116,7 +116,7 @@ arv_gvcp_packet_new_read_memory_ack (guint32 address, guint32 size, guint32 pack
  */
 
 ArvGvcpPacket *
-arv_gvcp_packet_new_write_memory_cmd (guint32 address, guint32 size, guint32 packet_count, size_t *packet_size)
+arv_gvcp_packet_new_write_memory_cmd (guint32 address, guint32 size, guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -149,7 +149,7 @@ arv_gvcp_packet_new_write_memory_cmd (guint32 address, guint32 size, guint32 pac
 
 ArvGvcpPacket *
 arv_gvcp_packet_new_write_memory_ack (guint32 address,
-				      guint32 packet_count, size_t *packet_size)
+				      guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -181,7 +181,7 @@ arv_gvcp_packet_new_write_memory_ack (guint32 address,
  */
 
 ArvGvcpPacket *
-arv_gvcp_packet_new_read_register_cmd (guint32 address, guint32 packet_count, size_t *packet_size)
+arv_gvcp_packet_new_read_register_cmd (guint32 address, guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -214,7 +214,7 @@ arv_gvcp_packet_new_read_register_cmd (guint32 address, guint32 packet_count, si
 
 ArvGvcpPacket *
 arv_gvcp_packet_new_read_register_ack (guint32 value,
-				       guint32 packet_count, size_t *packet_size)
+				       guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_value = g_htonl (value);
@@ -248,7 +248,7 @@ arv_gvcp_packet_new_read_register_ack (guint32 value,
 
 ArvGvcpPacket *
 arv_gvcp_packet_new_write_register_cmd (guint32 address, guint32 value,
-					guint32 packet_count, size_t *packet_size)
+					guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -283,7 +283,7 @@ arv_gvcp_packet_new_write_register_cmd (guint32 address, guint32 value,
 
 ArvGvcpPacket *
 arv_gvcp_packet_new_write_register_ack 	(guint32 address,
-					 guint32 packet_count, size_t *packet_size)
+					 guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 n_address = g_htonl (address);
@@ -373,7 +373,7 @@ arv_gvcp_packet_new_discovery_ack (size_t *packet_size)
 ArvGvcpPacket *
 arv_gvcp_packet_new_packet_resend_cmd (guint32 frame_id,
 				       guint32 first_block, guint32 last_block,
-				       guint32 packet_count, size_t *packet_size)
+				       guint16 packet_count, size_t *packet_size)
 {
 	ArvGvcpPacket *packet;
 	guint32 *data;
